rebase and fix OVBasicPropertiesTestsP SetConfigAllThrows test
