+++ src/tests/functional/plugin/shared/src/behavior/ov_plugin/properties_tests.cpp
@@ -135,7 +135,7 @@ TEST_P(OVBasicPropertiesTestsP, GetMetricThrowUnsupported) {
 }
 
 TEST_P(OVBasicPropertiesTestsP, SetConfigAllThrows) {
-    ov::Core ie;
+    ov::Core ie = ov::test::utils::create_core();
     OV_ASSERT_NO_THROW(ie.set_property({{"unsupported_key", "4"}}));
     ASSERT_ANY_THROW(ie.get_versions(target_device));
 }
